+++ includes/functions.php
@@ -23,6 +23,15 @@ use Activitypub\Transformer\Factory as Transformer_Factory;
 function get_context() {
 	$context = Activity::JSON_LD_CONTEXT;
 
+	/**
+	 * Filters the ActivityPub JSON-LD context.
+	 *
+	 * This filter allows developers to modify or extend the JSON-LD context used
+	 * in ActivityPub responses. The context defines the vocabulary and terms used
+	 * in the ActivityPub JSON objects.
+	 *
+	 * @param array $context The default ActivityPub JSON-LD context array.
+	 */
 	return \apply_filters( 'activitypub_json_context', $context );
 }
 
@@ -70,6 +79,16 @@ function get_webfinger_resource( $user_id ) {
  * @return array|WP_Error The Actor profile as array or WP_Error on failure.
  */
 function get_remote_metadata_by_actor( $actor, $cached = true ) {
+	/**
+	 * Filters the metadata before it is retrieved from a remote actor.
+	 *
+	 * Passing a non-false value will effectively short-circuit the remote request,
+	 * returning that value instead.
+	 *
+	 * @param mixed  $pre   The value to return instead of the remote metadata.
+	 *                      Default false to continue with the remote request.
+	 * @param string $actor The actor URL.
+	 */
 	$pre = apply_filters( 'pre_get_remote_metadata_by_actor', false, $actor );
 	if ( $pre ) {
 		return $pre;
@@ -416,7 +435,7 @@ function is_post_disabled( $post ) {
 		$disabled = true;
 	}
 
-	/*
+	/**
 	 * Allow plugins to disable posts for ActivityPub.
 	 *
 	 * @param boolean  $disabled True if the post is disabled, false otherwise.
@@ -603,7 +622,7 @@ function is_json( $data ) {
 }
 
 /**
- * Check whther a blog is public based on the `blog_public` option.
+ * Check whether a blog is public based on the `blog_public` option.
  *
  * @return bool True if public, false if not
  */
@@ -1054,7 +1073,7 @@ function get_enclosures( $post_id ) {
  *
  * @param int|\WP_Comment $comment Comment ID or comment object.
  *
- * @return \WP_Comment[] Array of ancestor comments or empty array if there are none.
+ * @return int[] Array of ancestor IDs.
  */
 function get_comment_ancestors( $comment ) {
 	$comment = \get_comment( $comment );
@@ -1314,11 +1333,7 @@ function generate_post_summary( $post, $length = 500 ) {
 	$content = \sanitize_post_field( 'post_excerpt', $post->post_excerpt, $post->ID );
 
 	if ( $content ) {
-		/**
-		 * Filters the post excerpt.
-		 *
-		 * @param string $content The post excerpt.
-		 */
+		/** This filter is documented in wp-includes/post-template.php */
 		return \apply_filters( 'the_excerpt', $content );
 	}
 
@@ -1356,6 +1371,7 @@ function generate_post_summary( $post, $length = 500 ) {
 
 	/*
 	Removed until this is merged: https://github.com/mastodon/mastodon/pull/28629
+	/** This filter is documented in wp-includes/post-template.php
 	return \apply_filters( 'the_excerpt', $content );
 	*/
 	return $content;
@@ -1461,6 +1477,15 @@ function get_content_visibility( $post_id ) {
 		$_visibility = $visibility;
 	}
 
+	/**
+	 * Filters the visibility of a post.
+	 *
+	 * @param string   $_visibility The visibility of the post. Possible values are:
+	 *                              - 'public': Post is public and federated.
+	 *                              - 'quiet_public': Post is public but not federated.
+	 *                              - 'local': Post is only visible locally.
+	 * @param \WP_Post $post        The post object.
+	 */
 	return \apply_filters( 'activitypub_content_visibility', $_visibility, $post );
 }
 
@@ -1514,7 +1539,7 @@ function get_upload_baseurl() {
 	/**
 	 * Filters the upload base URL.
 	 *
-	 * @param string \wp_get_upload_dir()['baseurl'] The upload base URL.
+	 * @param string $upload_dir The upload base URL. Default \wp_get_upload_dir()['baseurl']
 	 */
 	return apply_filters( 'activitypub_get_upload_baseurl', $upload_dir['baseurl'] );
 }